+++ plugins/module_utils/netbox_extras.py
@@ -13,6 +13,10 @@ from ansible_collections.netbox.netbox.plugins.module_utils.netbox_utils import 
 
 NB_CONFIG_CONTEXTS = "config_contexts"
 NB_TAGS = "tags"
+NB_CUSTOM_FIELDS = "custom_fields"
+NB_CUSTOM_LINKS = "custom_links"
+NB_EXPORT_TEMPLATES = "export_templates"
+NB_WEBHOOKS = "webhooks"
 
 
 class NetboxExtrasModule(NetboxModule):